+++ backend/src/homeassistant/generate.go
@@ -8,3 +8,5 @@ package homeassistant
 //go:generate go tool oapi-codegen -config cfg.yaml  -package mount -o mount/client.gen.go mount.yaml
 //go:generate go tool oapi-codegen -config cfg.yaml  -package ingress -o ingress/client.gen.go ingress.yaml
 //go:generate go tool oapi-codegen -config cfg.yaml  -package host -o host/client.gen.go host.yaml
+//go:generate go tool oapi-codegen -config cfg.yaml  -package addons -o addons/client.gen.go addons.yaml
+//go:generate go tool oapi-codegen -config cfg.yaml  -package resolution -o resolution/client.gen.go resolution.yaml