+++ CHANGELOG.md
@@ -5,9 +5,10 @@
 - [X] Manage WSDD2 service
 - [X] Manage Avahi service
 ###### __🚧 Work in progess__
-- [ ] Manage `local master`option
+- [ ] Manage `local master`option (?)
 - [ ] Veto files for share not global
 - [ ] Monitor tab
+- [ ] Help screen or overlay help/tour
 
 ###  🐛 Bug Fixes
 - [X] `enable`/`disable` share functionality is not working as expected.
@@ -19,6 +20,9 @@
 - [X] Implemet wachdog
 ###### __🚧 Work in progess__
 - [ ] Create the base documentation
+- [ ] Display version from ADDON
+- [ ] Align UI elements to HA
+    - [ ] Navbar h
 
 ## 2025.06.1-dev.801 [ 🧪 Pre-release ]