+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 2025.06.* [ 🚧 Unreleased ]
+## 2025.08.* [ 🚧 Unreleased ]
 
 ###  ✨ Features
 - Manage `recycle bin`option for share
@@ -11,12 +11,13 @@
 - [ ] Manage `local master`option (?)
 - [ ] Help screen or overlay help/tour [#82](https://github.com/dianlight/srat/issues/82)
 - [ ] Custom component [#83](https://github.com/dianlight/srat/issues/83)
-- [ ] Smart Control [#100](https://github.com/dianlight/srat/issues/100)
-- [ ] HDD Spin down [#101](https://github.com/dianlight/srat/issues/101)
+- [x] Smart Control [#100](https://github.com/dianlight/srat/issues/100)
+- [x] HDD Spin down [#101](https://github.com/dianlight/srat/issues/101)
 
 ###  🐛 Bug Fixes
 - `enable`/`disable` share functionality is not working as expected.
 - Renaming the admin user does not correctly create the new user or rename the existing one; issues persist until a full addon reboot.
+- Fix dianlight/hassio-addons#448 [SambaNAS2] Unable to create share for mounted volume
 #### __🚧 Work in progess__
 - [W] Addon protected mode check [#80](https://github.com/dianlight/srat/issues/85)